backend/requirements.txt
# === MagicMentor Backend Dependencies ===

# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12

# Data validation
pydantic==2.9.0
pydantic-settings==2.5.0

# AI — Perplexity (ALL agents: mentor, learning, matching, CV parsing, memory)
# sonar $1/$1 | sonar-pro $3/$15 | sonar-reasoning-pro $2/$8 per 1M tokens
openai>=1.40.0  # Perplexity uses OpenAI-compatible API

# Database
sqlalchemy==2.0.35

# CV Parsing
PyPDF2==3.0.1
pdfminer.six==20221105

# Job Scraping — multiple free sources
python-jobspy==1.1.80     # LinkedIn, Indeed, ZipRecruiter, Google Jobs

# Vector Memory (Tier 2 episodic memory — semantic search over past sessions)
chromadb>=0.5.0           # Vector store
sentence-transformers>=3.0.0  # Embeddings (all-MiniLM-L6-v2)

# AI Web Scraping (direct job site scraping fallback)
crawl4ai>=0.4.0           # LLM-powered web scraper

# Environment
python-dotenv==1.0.1

# Utils
httpx==0.27.0
